+++ ckanext/toolbelt/cli/__init__.py
@@ -4,6 +4,7 @@ from .make import make
 from .dev import dev
 from .search_index import search_index
 
+
 @click.group(short_help="Variety of useful commands.")
 def toolbelt():
     """Variety of useful commands."""