+++ ckanext/toolbelt/cli/search_index.py
@@ -9,16 +9,25 @@ from ckan.lib.search import query_for, clear
 def search_index():
     pass
 
+
 @search_index.command()
 @click.option("-y", "--no-confirm", is_flag=True)
 @click.pass_context
 def clear_missing(ctx, no_confirm: bool):
     q = query_for("package")
     with ctx.meta["flask_app"].test_request_context():
-        limit = tk.get_action("package_search")({"ignore_auth": True}, {"rows": 0})["count"]
+        limit = tk.get_action("package_search")(
+            {"ignore_auth": True}, {"rows": 0}
+        )["count"]
     with click.progressbar(q.get_all_entity_ids(limit)) as bar:
         query = model.Session.query(model.Package.id)
-        ids = {id for id in bar if not model.Session.query(query.filter_by(id=id).exists()).scalar()}
+        ids = {
+            id
+            for id in bar
+            if not model.Session.query(
+                query.filter_by(id=id).exists()
+            ).scalar()
+        }
 
     if not ids:
         click.secho("No missing packages detected", fg="green")
@@ -29,7 +38,9 @@ def clear_missing(ctx, no_confirm: bool):
         click.echo("\t" + id)
 
     if not no_confirm:
-        abort = not click.confirm("Do you want to remove these packages from the search index?")
+        abort = not click.confirm(
+            "Do you want to remove these packages from the search index?"
+        )
         if abort:
             raise click.Abort()