+++ ckanext/toolbelt/plugins/__init__.py
@@ -1,0 +1,5 @@
+import ckan.plugins as plugins
+
+
+class ToolbeltPlugin(plugins.SingletonPlugin):
+    pass