+++ ckanext/toolbelt/tests/test_decorators.py
@@ -2,8 +2,10 @@ import pytest
 
 from ckanext.toolbelt import decorators
 
+
 def fun():
     pass
+
 
 def test_availability():
     assert decorators.Collector()(fun)